refactor: change Stat to required
