+++ ui/jobs/components/ast.ts
@@ -2,8 +2,9 @@ import TimerBox from '../../../resources/timerbox';
 import { JobDetail } from '../../../types/event';
 import { kAbility } from '../constants';
 import { Bars } from '../jobs';
+import { Stats } from '../player';
 
-import { BaseComponent, Stats } from './base';
+import { BaseComponent } from './base';
 
 const cardsMap = {
   'Balance': { 'bonus': 'melee', 'seal': 'Solar' },
@@ -104,7 +105,7 @@ export default class AstComponent extends BaseComponent {
     }
   }
 
-  onStatChange(stat: Stats): void {
+  onStatChange(stat: Required<Stats>): void {
     this.combustBox.valuescale = stat.gcdSpell;
     this.combustBox.threshold = stat.gcdSpell + 1;
     this.drawBox.valuescale = stat.gcdSpell;
